# app/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: QuickEx Backend API
  description: >
    QuickEx is a Stellar-based payment platform enabling payment links,
    recurring subscriptions, username resolution, marketplace trading,
    fiat on/off ramps, and webhook-driven event notifications.

    **Base URL (production):** `https://api.quickex.example.com`
    **Base URL (development):** `http://localhost:3000`

    Interactive Swagger UI is available at `/docs` when the server is running.
  version: 1.0.0
  contact:
    name: QuickEx API Support
    url: https://github.com/quickex/quickex
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.quickex.example.com
    description: Production
...
security:
  - ApiKeyAuth: []
  - {}
paths:
...
  /fiat-ramps/anchors:
    get:
      tags: [fiat-ramps]
      operationId: getAnchors
      summary: Fetch available anchors
      parameters:
        - name: assetCode
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: anchorDomain
          in: query
          description: Optional configured anchor domain; otherwise configured anchors are discovered.
          schema:
            type: string
      responses:
        "200":
          description: Available anchors
  /fiat-ramps/sep10/challenge:
    get:
      tags: [fiat-ramps]
      operationId: getSep10Challenge
      summary: Fetch a SEP-10 transaction challenge for client-side signing
      parameters:
        - name: anchorDomain
          in: query
          required: true
          schema:
            type: string
        - name: userAccount
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Anchor-issued SEP-10 challenge transaction
  /fiat-ramps/deposit:
    post:
      tags: [fiat-ramps]
      operationId: initiateDeposit
      summary: Initiate SEP-24 hosted deposit flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FiatRampRequest"
      responses:
        "201":
          description: Deposit flow initiated
  /fiat-ramps/withdraw:
    post:
      tags: [fiat-ramps]
      operationId: initiateWithdrawal
      summary: Initiate SEP-24 hosted withdrawal flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FiatRampRequest"
      responses:
        "201":
          description: Withdrawal flow initiated
...
  /fiat-ramps/kyc/status:
    post:
      tags: [fiat-ramps]
      operationId: getKycStatus
      summary: Fetch normalized customer KYC status using SEP-10 authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [anchorDomain, userAccount, signedChallenge]
              properties:
                anchorDomain:
                  type: string
                userAccount:
                  type: string
                signedChallenge:
                  type: string
                customerId:
                  type: string
      responses:
        "201":
          description: Normalized SEP-12 KYC status
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for higher rate limits

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
...
  schemas:
...
    FiatRampRequest:
      type: object
      required: [assetCode, amount, userAccount, anchorDomain, signedChallenge]
      properties:
        assetCode:
          type: string
        amount:
          oneOf:
            - type: number
            - type: string
          description: Positive amount, preferably a decimal string to preserve precision.
        userAccount:
          type: string
        anchorDomain:
          type: string
        signedChallenge:
          type: string
          description: SEP-10 challenge XDR signed by the account's wallet; the backend never receives a secret key.
        countryCode:
          type: string
        lang:
          type: string

    # --- Refunds ---
...